add upper menu picker
